Replaced anchor tag with React-Router's Link component.
Anchors were used for creating menu and links in the header.
Because of that whenever menu or link in the header was clicked,
browser was making a GET request to the server. Since server
didn't had the said resource, as a fallback mechanism it was
sending index.html. That means the app was making a complete
roundtrip for each route defined and page was completely
reloaded.

React Router provides a <Link> component to prevent that from
happening. When clicking a <Link>, the URL will be updated and
the rendered content will change without reloading the page.